Improve dual chn implementation. Upgrade to c++17 standard

+++ nacs-spcm/Stream.h
@@ -104,7 +104,7 @@ public:
     }
     static Cmd getAddChn(uint32_t t, uint32_t chn)
     {
-        return Cmd{t, (uint8_t)CmdType::ModChn, add_chn, chn}; //overload NOT meant to be used in stream. Meant for usage with real chn ID not chn within a stream
+        return Cmd{t, (uint8_t)CmdType::ModChn, add_chn, static_cast<int32_t>(chn)}; //overload NOT meant to be used in stream. Meant for usage with real chn ID not chn within a stream
     }
     static Cmd getDelChn(uint32_t t, uint32_t chn)
     {
